+++ plugins/destination/plugin_testing_write_append.go
@@ -20,9 +20,13 @@ func (s *PluginTestSuite) destinationPluginTestWriteAppend(ctx context.Context, 
 	}
 	tableName := spec.Name
 	table := testdata.TestSourceSchema(tableName, testdata.TestSourceOptions{
-		IncludeDates:   false,
-		IncludeMaps:    false,
-		IncludeStructs: false,
+		IncludeDates:      false,
+		IncludeMaps:       false,
+		IncludeStructs:    false,
+		IncludeIntervals:  false,
+		IncludeDurations:  false,
+		IncludeTimes:      false,
+		IncludeLargeTypes: false,
 	})
 	syncTime := time.Now().UTC().Round(1 * time.Second)
 	tables := []*arrow.Schema{
