+++ testdata/table.go
@@ -1,0 +1,121 @@
+package testdata
+
+import "github.com/cloudquery/plugin-sdk/v2/schema"
+
+// TestSourceTable returns a test table with the given name.
+// Deprecated: Use TestSourceSchema instead.
+func TestSourceTable(name string) *schema.Table {
+	return &schema.Table{
+		Name:        name,
+		Description: "Test table",
+		Columns: schema.ColumnList{
+			schema.CqIDColumn,
+			schema.CqParentIDColumn,
+			{
+				Name:            "uuid_pk",
+				Type:            schema.TypeUUID,
+				CreationOptions: schema.ColumnCreationOptions{PrimaryKey: true},
+			},
+			{
+				Name:            "string_pk",
+				Type:            schema.TypeString,
+				CreationOptions: schema.ColumnCreationOptions{PrimaryKey: true},
+			},
+			{
+				Name: "bool",
+				Type: schema.TypeBool,
+			},
+			{
+				Name: "int",
+				Type: schema.TypeInt,
+			},
+			{
+				Name: "float",
+				Type: schema.TypeFloat,
+			},
+			{
+				Name: "uuid",
+				Type: schema.TypeUUID,
+			},
+			{
+				Name: "text",
+				Type: schema.TypeString,
+			},
+			{
+				Name: "text_with_null",
+				Type: schema.TypeString,
+			},
+			{
+				Name: "bytea",
+				Type: schema.TypeByteArray,
+			},
+			{
+				Name: "text_array",
+				Type: schema.TypeStringArray,
+			},
+			{
+				Name: "text_array_with_null",
+				Type: schema.TypeStringArray,
+			},
+			{
+				Name: "int_array",
+				Type: schema.TypeIntArray,
+			},
+			{
+				Name: "timestamp",
+				Type: schema.TypeTimestamp,
+			},
+			{
+				Name: "json",
+				Type: schema.TypeJSON,
+			},
+			{
+				Name: "uuid_array",
+				Type: schema.TypeUUIDArray,
+			},
+			{
+				Name: "inet",
+				Type: schema.TypeInet,
+			},
+			{
+				Name: "inet_array",
+				Type: schema.TypeInetArray,
+			},
+			{
+				Name: "cidr",
+				Type: schema.TypeCIDR,
+			},
+			{
+				Name: "cidr_array",
+				Type: schema.TypeCIDRArray,
+			},
+			{
+				Name: "macaddr",
+				Type: schema.TypeMacAddr,
+			},
+			{
+				Name: "macaddr_array",
+				Type: schema.TypeMacAddrArray,
+			},
+		},
+	}
+}
+
+// TestTableIncremental returns an incremental test table.
+// Deprecated: Use TestSourceSchemaWithMetadata instead.
+func TestTableIncremental(name string) *schema.Table {
+	t := TestTable(name)
+	t.IsIncremental = true
+	return t
+}
+
+// TestTable returns a table with columns of all CQ types. Useful for destination testing purposes.
+// Deprecated: Use TestSourceSchema instead.
+func TestTable(name string) *schema.Table {
+	sourceTable := TestSourceTable(name)
+	sourceTable.Columns = append(schema.ColumnList{
+		schema.CqSourceNameColumn,
+		schema.CqSyncTimeColumn,
+	}, sourceTable.Columns...)
+	return sourceTable
+}
